requirements.txt
numpy==1.26.4
requests==2.32.3
boto3==1.35.3
botocore==1.35.3
python-rapidjson==1.20
pytest-dependency==0.6.0
prettytable==3.11.0
PyYAML==6.0.1
html2text==2024.2.26
lxml==5.3.0
python-dateutil==2.8.2
cpe==1.3.0
packageurl-python==0.15.6
pandas==2.0.3
psycopg2-binary==2.9.10
openpyxl==3.1.5
duckdb==1.2.0